requirements.txt
streamlit
langchain_core
langchain-community
python-dotenv
langserve
fastapi
uvicorn
sse_starlette